# services/immersion-api/http/rest/openapi/api.yaml
openapi: 3.0.3
info:
  title: immersion-api
  license:
    name: MIT
    url: https://github.com/tadoku/tadoku/blob/main/LICENSE
  version: 2.0.0
...
servers:
  - url: https://tadoku.app/api/immersion/
paths:
...
  /logs:
    post:
      summary: Submits a new log
      operationId: logCreate
      tags: [logs]
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - registration_ids
                - language_code
                - activity_id
                - amount
                - unit_id
                - tags
              properties:
                registration_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                language_code:
                  type: string
                  example: zho
                activity_id:
                  format: int32
                  type: integer
                amount:
                  type: number
                  format: float
                unit_id:
                  type: string
                  format: uuid
                tags:
                  type: array
                  items:
                    type: string
                  example: ['book', 'fiction']
                description:
                  type: string
      responses:
        '200':
          description: successful operation
        '400':
          description: invalid submission
  /logs/{id}:
    get:
      summary: Fetches a log by id
      operationId: logFindByID
      tags: [logs]
      parameters:
        - name: id
          in: path
          description: ID of log to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Log'
        '404':
          description: not found
  /logs/configuration-options:
    get:
      summary: Fetches the configuration options for a log
      operationId: logGetConfigurations
      tags: [logs]
      security:
        - cookieAuth: []
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogConfigurationOptions'
  /users/{userId}/profile:
    get:
      summary: Fetches a profile of a user
      operationId: profileFindByUserID
      tags: [profile]
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '404':
          description: not found
...
  /users/{userId}/contest-registrations/{year}:
    get:
      summary: Fetches the contest registrations of a user for a given year
      description: Includes private contests for the current user and admins
      operationId: profileYearlyContestRegistrationsByUserID
      tags: [profile]
      parameters:
        - name: userId
          in: path
          description: ID of user to return
          required: true
          schema:
            type: string
            format: uuid
        - name: year
          in: path
          description: year of activity to fetch
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContestRegistrations'
...
components:
  schemas:
    UserProfile:
      type: object
      required:
        - id
        - display_name
        - created_at
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
          example: 2022-12-14T19:48:00Z
...
    LogConfigurationOptions:
      allOf:
        - $ref: "#/components/schemas/Languages"
        - $ref: "#/components/schemas/Activities"
        - $ref: "#/components/schemas/Units"
        - $ref: "#/components/schemas/Tags"
    Unit:
      type: object
      required:
        - id
        - log_activity_id
        - name
        - modifier
      properties:
        id:
          type: string
          format: uuid
        log_activity_id:
          type: integer
        name:
          type: string
          example: Page
        modifier:
          type: number
          example: 3.14
        language_code:
          type: string
          example: jpa
    Units:
      type: object
      required:
        - units
      properties:
        units:
          type: array
          items:
            $ref: "#/components/schemas/Unit"
    Tag:
      type: object
      required:
        - id
        - log_activity_id
        - name
      properties:
        id:
          type: string
          format: uuid
        log_activity_id:
          type: integer
        name:
          type: string
          example: Book
    Tags:
      type: object
      required:
        - tags
      properties:
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
...
    Language:
      type: object
      required:
        - code
        - name
      properties:
        code:
          type: string
          example: jpa
          description: In ISO-639-3 https://en.wikipedia.org/wiki/Wikipedia:WikiProject_Languages/List_of_ISO_639-3_language_codes_(2019)
        name:
          type: string
          example: Japanese
    Languages:
      type: object
      required:
        - languages
      properties:
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
    Activity:
      type: object
      required:
        - id
        - name
      properties:
        id:
          format: int32
          type: integer
          example: 1
        name:
          type: string
          example: Reading
        default:
          type: boolean
    Activities:
      type: object
      required:
        - activities
      properties:
        activities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
    ContestBase:
      type: object
      required:
        - contest_start
        - contest_end
        - registration_end
        - title
        - private
        - official
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
          example: 2023 Round 1
        description:
          type: string
        contest_start:
          type: string
          format: date
        contest_end:
          type: string
          format: date
        registration_end:
          type: string
          format: date
        private:
          type: boolean
        official:
          type: boolean
          example: false
        owner_user_id:
          type: string
          format: uuid
        owner_user_display_name:
          type: string
          example: "john"
        created_at:
          type: string
          format: date-time
          example: 2022-12-14T19:48:00Z
        updated_at:
          type: string
          format: date-time
          example: 2022-12-14T19:48:00Z
        deleted:
          type: boolean
...
    ContestView:
      allOf:
        - $ref: '#/components/schemas/ContestBase'
        - type: object
          required:
            - allowed_languages
            - allowed_activities
          properties:
            allowed_languages:
              type: array
              items:
                $ref: "#/components/schemas/Language"
            allowed_activities:
              type: array
              items:
                $ref: "#/components/schemas/Activity"
...
    ContestRegistration:
      type: object
      required:
        - contest_id
        - user_id
        - user_display_name
        - languages
      properties:
        id:
          type: string
          format: uuid
        contest_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_display_name:
          type: string
          example: "john"
        languages:
          type: array
          items:
            $ref: "#/components/schemas/Language"
        contest:
          $ref: "#/components/schemas/ContestView"
    ContestRegistrations:
      allOf:
        - $ref: '#/components/schemas/PaginatedList'
        - type: object
          required:
            - registrations
          properties:
            registrations:
              type: array
              maxItems: 50
              items:
                $ref: "#/components/schemas/ContestRegistration"
    ContestRegistrationReference:
      type: object
      required:
        - contest_id
        - registration_id
        - title
      properties:
        registration_id:
          type: string
          format: uuid
        contest_id:
          type: string
          format: uuid
        title:
          type: string
          example: "January 2026"
...
    Log:
      type: object
      required:
        - id
        - user_id
        - language
        - activity
        - unit_name
        - tags
        - amount
        - modifier
        - score
        - created_at
        - deleted
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        user_display_name:
          type: string
        description:
          type: string
        language:
          $ref: "#/components/schemas/Language"
        activity:
          $ref: "#/components/schemas/Activity"
        unit_name:
          type: string
          example: "page"
        tags:
          type: array
          items:
            type: string
          example: ['Book', 'Fiction']
        amount:
          type: number
          format: float
        modifier:
          type: number
          format: float
        score:
          type: number
          format: float
        registrations:
          type: array
          items:
            $ref: "#/components/schemas/ContestRegistrationReference"
        created_at:
          type: string
          format: date-time
        deleted:
          type: boolean
...
    PaginatedList:
      type: object
      required:
        - next_page_token
        - total_size
      properties:
        total_size:
          type: integer
          example: 100
        next_page_token:
          type: string
          description: is empty if there's no next page
          example: "3"
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: ory_kratos_session